implement custom exception handling

+++ src/main/java/com/MarianFinweFeanor/Virtual_Teacher/Service/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.MarianFinweFeanor.Virtual_Teacher.Service;
 
 import com.MarianFinweFeanor.Virtual_Teacher.Model.UserRole;
+import com.MarianFinweFeanor.Virtual_Teacher.exceptions.EntityDuplicateException;
+import com.MarianFinweFeanor.Virtual_Teacher.exceptions.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.MarianFinweFeanor.Virtual_Teacher.Model.User;
@@ -24,7 +26,7 @@ public class UserServiceImpl implements UserService {
     public User saveUser(User user)
     {
         if (userRepository.existsByEmail(user.getEmail())) {
-            throw new RuntimeException("Email already in use!");
+            throw new EntityDuplicateException("User", "email", user.getEmail());
         }
 
         // If `role` is null, assign default role as STUDENT
@@ -38,18 +40,28 @@ public class UserServiceImpl implements UserService {
     // Get all users
     @Override
     public List<User> getAllUsers() {
+        if(userRepository.count() == 0) {
+            throw new EntityNotFoundException("Users","database");
+        }
+
         return userRepository.findAll();
     }
 
     // Get a user by ID
     @Override
     public Optional<User> getUserById(Long userId) {
+        if(!userRepository.existsById(userId)) {
+            throw new EntityNotFoundException("User", userId);
+        }
         return userRepository.findById(userId);
     }
 
     // Delete a user by ID
     @Override
     public void deleteUserById(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new EntityNotFoundException("User", userId);
+        }
         userRepository.deleteById(userId);
     }
